+++ packages/railway/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @computesdk/railway
 
+## 1.0.2
+
+### Patch Changes
+
+- a146b97: Adding in proper background command via background: true
+- Updated dependencies [a146b97]
+  - computesdk@1.8.8
+
 ## 1.0.1
 
 ### Patch Changes